% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{Client}
\alias{Client}
\title{Client for the Vantage federated learning infrastructure}
\usage{
Client(host, username, password, collaboration_id, api_path = "")
}
\arguments{
\item{host}{hostname (may include port, separated using a colon, e.g. 'https://localhost:8000')}

\item{username}{username}

\item{password}{password}

\item{collaboration_id}{collaboration ID to post tasks to}

\item{api_path}{alternative base path, is appended to host}
}
\value{
Client
}
\description{
Simplifies communication with the REST-api by handling authentication and
offering methods for GET/PUT/POST requests.
}
\examples{
client <- Client(
    'https://api-test.distributedlearning.ai',
    'admin@example.com',
    'password',
    1
)

tasks <- client$GET("/task")
}
