% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.client.R
\docType{class}
\name{ContainerClient}
\alias{ContainerClient}
\title{Client for master container algorithms on the vantage6 infrastructure.}
\description{
Client for master container algorithms on the vantage6 infrastructure.

Client for master container algorithms on the vantage6 infrastructure.
}
\section{Super class}{
\code{\link[vtg:Client]{vtg::Client}} -> \code{ContainerClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ContainerClient$new()}}
\item \href{#method-setUseEncryption}{\code{ContainerClient$setUseEncryption()}}
\item \href{#method-clone}{\code{ContainerClient$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="GET">}\href{../../vtg/html/Client.html#method-GET}{\code{vtg::Client$GET()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="POST">}\href{../../vtg/html/Client.html#method-POST}{\code{vtg::Client$POST()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="PUT">}\href{../../vtg/html/Client.html#method-PUT}{\code{vtg::Client$PUT()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="authenticate">}\href{../../vtg/html/Client.html#method-authenticate}{\code{vtg::Client$authenticate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="call">}\href{../../vtg/html/Client.html#method-call}{\code{vtg::Client$call()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="decrypt.result">}\href{../../vtg/html/Client.html#method-decrypt.result}{\code{vtg::Client$decrypt.result()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="encrypt">}\href{../../vtg/html/Client.html#method-encrypt}{\code{vtg::Client$encrypt()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="getCollaboration">}\href{../../vtg/html/Client.html#method-getCollaboration}{\code{vtg::Client$getCollaboration()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="getCollaborations">}\href{../../vtg/html/Client.html#method-getCollaborations}{\code{vtg::Client$getCollaborations()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="getOrganization">}\href{../../vtg/html/Client.html#method-getOrganization}{\code{vtg::Client$getOrganization()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="getVersion">}\href{../../vtg/html/Client.html#method-getVersion}{\code{vtg::Client$getVersion()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="process.results">}\href{../../vtg/html/Client.html#method-process.results}{\code{vtg::Client$process.results()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="refresh.token">}\href{../../vtg/html/Client.html#method-refresh.token}{\code{vtg::Client$refresh.token()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="repr">}\href{../../vtg/html/Client.html#method-repr}{\code{vtg::Client$repr()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="request">}\href{../../vtg/html/Client.html#method-request}{\code{vtg::Client$request()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="set.task.image">}\href{../../vtg/html/Client.html#method-set.task.image}{\code{vtg::Client$set.task.image()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="setCollaborationId">}\href{../../vtg/html/Client.html#method-setCollaborationId}{\code{vtg::Client$setCollaborationId()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="setPrivateKey">}\href{../../vtg/html/Client.html#method-setPrivateKey}{\code{vtg::Client$setPrivateKey()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="setUseMasterContainer">}\href{../../vtg/html/Client.html#method-setUseMasterContainer}{\code{vtg::Client$setUseMasterContainer()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vtg" data-topic="Client" data-id="wait.for.results">}\href{../../vtg/html/Client.html#method-wait.for.results}{\code{vtg::Client$wait.for.results()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContainerClient$new(host, token, api_path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setUseEncryption"></a>}}
\subsection{Method \code{setUseEncryption()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContainerClient$setUseEncryption(flag)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContainerClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
