% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.task.input.unserialized.R
\name{create.task.input.unserialized}
\alias{create.task.input.unserialized}
\title{Create a data structure that can be used as input for \code{\link[=dispatch.RPC]{dispatch.RPC()}}).}
\usage{
create.task.input.unserialized(use.master.container, method, ...)
}
\arguments{
\item{use.master.container}{logical Indicates whether the task should be run in a
master container.}

\item{method}{Character vector (string) of the method to run.}

\item{...}{Arguments and keyword arguments}
}
\value{
\code{\link[=list]{list()}} with keys \code{master}, \code{debug}, \code{method}, \code{args}, \code{kwargs}
}
\description{
Creates a \code{\link[base:list]{base::list()}} with keys \code{master}, \code{debug}, \code{method}, \code{args}
and \code{kwargs} that can be used by \code{\link[=dispatch.RPC]{dispatch.RPC()}}, run on a node, to
execute a remote procedure call.
}
