% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.RPC.R
\name{dispatch.RPC}
\alias{dispatch.RPC}
\title{Dispatch a remote procedure call.}
\usage{
dispatch.RPC(df, input)
}
\arguments{
\item{df}{data frame containing the \emph{local} dataset}

\item{input}{string containing serialized JSON; JSON should contain
the keys \code{method}, \code{args} and \code{kwargs}.}
}
\value{
Output depends on RPC call.
}
\description{
Dispatch a remote procedure call.
}
