% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.wrapper.R
\name{docker.wrapper}
\alias{docker.wrapper}
\title{Entrypoint when excecuting this script using Rscript within a Docker container.}
\usage{
docker.wrapper()
}
\value{
\code{null}; output will be written to \code{output.txt}
}
\description{
This function expects the environment variable \code{DATABASE_URI} to be set and
pointing towards a CSV file that can be read by the R-function \code{read.csv()}
\emph{without} specifying format-details.
}
\details{
Performs the following actions:
\itemize{
\item Load data from the CSV file
\item Loads the input parameters provided to the container in \code{input.txt}.
\item Wraps the docker input/output for \code{dispatch.RPC()}.
}

(de)serialization is performed in \code{dispatch.RPC()} to enable testing.
}
