% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.client.R
\name{MockClient}
\alias{MockClient}
\title{MockClient}
\description{
Client that mocks running an algorithm on the vantage6 infrastructure.
Particularly useful for testing and developing.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MockClient$new()}}
\item \href{#method-authenticate}{\code{MockClient$authenticate()}}
\item \href{#method-set.pkgname}{\code{MockClient$set.pkgname()}}
\item \href{#method-set.task.image}{\code{MockClient$set.task.image()}}
\item \href{#method-call}{\code{MockClient$call()}}
\item \href{#method-process.results}{\code{MockClient$process.results()}}
\item \href{#method-clone}{\code{MockClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockClient$new(datasets, pkgname)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-authenticate"></a>}}
\subsection{Method \code{authenticate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockClient$authenticate(username = "", password = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set.pkgname"></a>}}
\subsection{Method \code{set.pkgname()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockClient$set.pkgname(pkg)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set.task.image"></a>}}
\subsection{Method \code{set.task.image()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockClient$set.task.image(image, task.name = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-call"></a>}}
\subsection{Method \code{call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockClient$call(method, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process.results"></a>}}
\subsection{Method \code{process.results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockClient$process.results(site_results)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
