% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{class}
\name{Client}
\alias{Client}
\title{Client for the vantage6 infrastructure.}
\description{
Facilitates querying collaborations, creating tasks and retrieving
results.
}
\examples{
client <- vtg::Client$new('http://localhost:5000', api_path='/api')
client$authenticate('root', 'password')

collaborations <- client$getCollaborations()
print(collaborations)

client$set.task.image(image.name, task.name="colnames")
result <- client$call("colnames")


}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Client$new()}}
\item \href{#method-authenticate}{\code{Client$authenticate()}}
\item \href{#method-setPrivateKey}{\code{Client$setPrivateKey()}}
\item \href{#method-getVersion}{\code{Client$getVersion()}}
\item \href{#method-getCollaborations}{\code{Client$getCollaborations()}}
\item \href{#method-getOrganization}{\code{Client$getOrganization()}}
\item \href{#method-getCollaboration}{\code{Client$getCollaboration()}}
\item \href{#method-setCollaborationId}{\code{Client$setCollaborationId()}}
\item \href{#method-setUseEncryption}{\code{Client$setUseEncryption()}}
\item \href{#method-setUseMasterContainer}{\code{Client$setUseMasterContainer()}}
\item \href{#method-refresh.token}{\code{Client$refresh.token()}}
\item \href{#method-request}{\code{Client$request()}}
\item \href{#method-GET}{\code{Client$GET()}}
\item \href{#method-POST}{\code{Client$POST()}}
\item \href{#method-PUT}{\code{Client$PUT()}}
\item \href{#method-wait.for.results}{\code{Client$wait.for.results()}}
\item \href{#method-decrypt.result}{\code{Client$decrypt.result()}}
\item \href{#method-process.results}{\code{Client$process.results()}}
\item \href{#method-set.task.image}{\code{Client$set.task.image()}}
\item \href{#method-encrypt}{\code{Client$encrypt()}}
\item \href{#method-call}{\code{Client$call()}}
\item \href{#method-repr}{\code{Client$repr()}}
\item \href{#method-clone}{\code{Client$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$new(
  host,
  username = "",
  password = "",
  collaboration_id = NULL,
  api_path = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{host}}{\code{\link[=character]{character()}} Host to connect to}

\item{\code{username}}{\code{\link[=character]{character()}} (optional) Username}

\item{\code{password}}{\code{\link[=character]{character()}} (optional) Password}

\item{\code{collaboration_id}}{\code{\link[=integer]{integer()}} (optional) Collaboration id}

\item{\code{api_path}}{\code{\link[=character]{character()}} (optional) API path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-authenticate"></a>}}
\subsection{Method \code{authenticate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$authenticate(username = "", password = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{character (optional) Username. If not provided, the
username and password provided to the constructor are used.}

\item{\code{password}}{character (optional) Password. Required if username
is provided.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPrivateKey"></a>}}
\subsection{Method \code{setPrivateKey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$setPrivateKey(bytes_or_filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getVersion"></a>}}
\subsection{Method \code{getVersion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$getVersion()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCollaborations"></a>}}
\subsection{Method \code{getCollaborations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$getCollaborations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getOrganization"></a>}}
\subsection{Method \code{getOrganization()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$getOrganization(organization_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCollaboration"></a>}}
\subsection{Method \code{getCollaboration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$getCollaboration(collaboration_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setCollaborationId"></a>}}
\subsection{Method \code{setCollaborationId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$setCollaborationId(collaboration_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setUseEncryption"></a>}}
\subsection{Method \code{setUseEncryption()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$setUseEncryption(flag)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setUseMasterContainer"></a>}}
\subsection{Method \code{setUseMasterContainer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$setUseMasterContainer(flag = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-refresh.token"></a>}}
\subsection{Method \code{refresh.token()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$refresh.token()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-request"></a>}}
\subsection{Method \code{request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$request(method, path, data = NULL, first_try = T, prefix.api.path = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GET"></a>}}
\subsection{Method \code{GET()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$GET(path, prefix.api.path = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-POST"></a>}}
\subsection{Method \code{POST()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$POST(path, data = NULL, prefix.api.path = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PUT"></a>}}
\subsection{Method \code{PUT()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$PUT(path, data = NULL, prefix.api.path = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wait.for.results"></a>}}
\subsection{Method \code{wait.for.results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$wait.for.results(task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decrypt.result"></a>}}
\subsection{Method \code{decrypt.result()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$decrypt.result(serialized.output)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process.results"></a>}}
\subsection{Method \code{process.results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$process.results(site_results)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set.task.image"></a>}}
\subsection{Method \code{set.task.image()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$set.task.image(image, task.name = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-encrypt"></a>}}
\subsection{Method \code{encrypt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$encrypt(data, org)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-call"></a>}}
\subsection{Method \code{call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$call(method, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\subsection{Method \code{repr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$repr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
