% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.RPC.R
\name{dispatch.RPC}
\alias{dispatch.RPC}
\title{Dispatch a remote procedure call.}
\usage{
dispatch.RPC(df, input, pkg = "", token = "")
}
\arguments{
\item{df}{\code{\link[=data.frame]{data.frame()}} containing the \emph{local} dataset}

\item{input}{\code{\link[=list]{list()}} with at least the keys \code{method}, \code{args} and \code{kwargs}.
Key \code{debug} (\code{\link[=logical]{logical()}}) is optional.}
}
\value{
Output depends on RPC call.
}
\description{
This method is called by \code{\link[=docker.wrapper]{docker.wrapper()}}, supplying \code{df}, \code{input} and
\code{pkg}. Token will be supplied only when running within a master container.
}
